Move alignment failure error reporting to machine
